In order to support responsive design, I had to change up some components and move things around. This work will make it easier to shift smoothly between web and mobile.
Abstracted the menu component to just take children, and moved all upgrade logic into the StatUpgradeButton component.

Moved "close on click outside" logic into the menu to let me reuse it elsewhere.

Created a CharacterSheet component to make CharacterView less chonky. Will be further decomposing CharacterSheet into modules in later commits.

Added a ChipButton component and removed excess logic from ChipCountersContainer

Updated SpendBountyPointsPanel to accommodate some of these changes.

Added first Context in CharacterContext. Will be adding more in the future to reduce prop drilling.
