+++ src/contexts/characterContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const CharacterContext = createContext(1);
